feat - Cobro finalizado
