+++ components/screens/Pago.js
@@ -142,6 +142,7 @@ function Pago(props) {
 
             //console.log("response", links);
             setPaypalUrl(approvalUrl);
+            navigation.popToTop();
           })
           .catch(err => {
             //console.log({ ...err });
