+++ components/screens/Profile/ShowWalk.js
@@ -70,8 +70,8 @@ function ShowWalk(props) {
 
   const confirmFinishRequest = () => {
     Alert.alert(
-      "Finalizar Servicio",
-      "¿Estás seguro?",
+      "Va a navegar hacia una página para poner su email para obtener el cobro",
+      "¿Está seguro?",
       [
         {
           text: "Si",
@@ -86,35 +86,15 @@ function ShowWalk(props) {
     );
   };
 
-  const confirmAcceptPayment = () => {
-    Alert.alert(
-      "Acceder a la pasarela de pago",
-      "¿Estás seguro?",
-      [
-        {
-          text: "Si",
-          onPress: goToPayment,
-        },
-        {
-          text: "No",
-          style: "cancel",
-        },
-      ],
-      { cancelable: false }
-    );
-  };
 
   const finishRequest = () => {
-    var idRequest = request.id;
+    /*var idRequest = request.id;
     var query = db.ref().child("requests/" + idRequest);
     query.update({
       isFinished: true,
     });
     alert("Se ha finalizado el servicio correctamente");
-    navigation.popToTop();
-  };
-
-  const goToPayment = () => {
+    navigation.popToTop();*/
     navigation.navigate("Pagar", { request: request });
   };
 
@@ -309,9 +289,6 @@ function ShowWalk(props) {
               buttonStyle={requestsStyles.requestsBtn2}
               containerStyle={requestsStyles.requestsBtnContainer2}
               title="Finalizar Servicio"
-              buttonStyle={globalStyles.showWalkBtn3}
-              containerStyle={globalStyles.showWalkBtnContainer3}
-              title="Finalizar Servicio"
               onPress={confirmFinishRequest}
               icon={
                 <Icon
@@ -405,25 +382,6 @@ function ShowWalk(props) {
             </View>
           </View>
         </View>
-        <Button
-          buttonStyle={requestsStyles.requestsBtn2}
-          containerStyle={requestsStyles.requestsBtnContainer2}
-          title="Finalizar Servicio"
-          buttonStyle={globalStyles.showWalkBtn3}
-          containerStyle={globalStyles.showWalkBtnContainer3}
-          title="Pasar a la pasarela de cobro"
-          onPress={confirmAcceptPayment}
-          icon={
-            <Icon
-              type="font-awesome"
-              name="thumbs-up"
-              size={30}
-              color="white"
-              marginLeft={"10%"}
-            />
-          }
-          titleStyle={requestsStyles.requestsBtnTittle}
-        />
       </SafeAreaView>
     );
   } else {
